sign in button tweaks
