+++ apps/dotcom/client/public/tla/locales-compiled/en.json
@@ -58,7 +58,6 @@
   "d4bbf74f3e": "Toggle sidebar",
   "d6381957f5": "Share this project",
   "d721757161": "Theme",
-  "d9776f0775": "Sign up",
   "d9fe660e4e": "Sorry, you don’t have access to that workspace.",
   "dfdb78a4e4": "App debug flags",
   "e2cd8064c2": "Sorry, you don’t have access to that user.",
