+++ apps/dotcom/client/public/tla/locales/en.json
@@ -177,9 +177,6 @@
   "d721757161": {
     "translation": "Theme"
   },
-  "d9776f0775": {
-    "translation": "Sign up"
-  },
   "d9fe660e4e": {
     "translation": "Sorry, you don’t have access to that workspace."
   },
